functions/requirements.txt
firebase-functions
firebase-admin
fastapi
sqlalchemy
psycopg2-binary
pydantic[email]
passlib[bcrypt]
pyjwt
requests
numpy
pandas
python-multipart